+++ packages/python-packages/api-stub-generator/apistub/nodes/_property_node.py
@@ -15,7 +15,6 @@ class PropertyNode(NodeEntityBase):
         self.obj = obj
         self.read_only = True
         self.type = None
-        self.errors = []
         self.name = name
         self._inspect()
         # Generate ID using name found by inspect
@@ -44,7 +43,7 @@ class PropertyNode(NodeEntityBase):
                     if not self.type:
                         self.type = docstring_parser.ret_type
                 except:
-                    self.errors.append("Failed to find type of property {}".format(self.name))
+                    pass
 
         self.display_name = "{0}: {1}".format(self.name, self.type)
         if self.read_only:
@@ -57,17 +56,12 @@ class PropertyNode(NodeEntityBase):
         apiview.add_keyword("property")
         apiview.add_space()
         apiview.add_line_marker(self.namespace_id)
-        apiview.add_text(self.namespace_id, self.name)
+        apiview.add_text(self.name)
         apiview.add_punctuation(":")
         apiview.add_space()
-        apiview.add_type(self.type)  # TODO: Pass navigation ID if it is internal type
+        apiview.add_type(self.type)
         if self.read_only:
             apiview.add_whitespace(count=5)
             apiview.add_literal("# Read-only")
-
-
-    def print_errors(self):
-        if self.errors:
-            print("property: {}".format(self.name))
-            for e in self.errors:
-                print("    {}".format(e))
+        for err in self.pylint_errors:
+            err.generate_tokens(apiview, self.namespace_id)