+++ packages/python-packages/api-stub-generator/tests/apiview_test.py
@@ -5,6 +5,7 @@
 # --------------------------------------------------------------------------
 
 from apistub import ApiView, TokenKind, StubGenerator
+from apistub.nodes import PylintParser
 import os
 import tempfile
 
@@ -13,9 +14,9 @@ from ._test_util import _check, _render_string, _tokenize
 class StubGenTestArgs:
     pkg_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', 'apistubgentest'))
     temp_path = tempfile.gettempdir()
+    source_url = None
     out_path = None
     mapping_path = None
-    hide_report = None
     verbose = None
     filter_namespace = None
 
@@ -32,7 +33,7 @@ class TestApiView:
 
     def test_multiple_newline_only_add_one(self):
         apiview = ApiView()
-        apiview.add_text(None, "Something")
+        apiview.add_text("Something")
         apiview.add_newline()
         # subsequent calls result in no change
         apiview.add_newline()
@@ -44,7 +45,7 @@ class TestApiView:
         apiview.set_blank_lines(3)
         assert self._count_newlines(apiview) == 4 # +1 for carriage return
 
-        apiview.add_text(None, "Something")
+        apiview.add_text("Something")
         apiview.add_newline()
         apiview.set_blank_lines(1)
         apiview.set_blank_lines(5)
@@ -54,9 +55,11 @@ class TestApiView:
 
     def test_api_view_diagnostic_warnings(self):
         args = StubGenTestArgs()
-        print(args.pkg_path)
         stub_gen = StubGenerator(args=args)
         apiview = stub_gen.generate_tokens()
         # ensure we have only the expected diagnostics when testing apistubgentest
-        # TODO: This will be removed soon.
-        assert True
+        unclaimed = PylintParser.get_unclaimed()
+        assert len(apiview.diagnostics) == 4
+        # The "needs copyright header" error corresponds to a file, which isn't directly
+        # represented in APIView
+        assert len(unclaimed) == 1